Split tests, cover validation
